requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
click==8.1.3
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.23.2
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
